% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armmMod_class.R
\name{fixef.armmMod}
\alias{fixef.armmMod}
\title{Extract fixed-effects estimates from an \code{armmMod} object}
\usage{
fixef.armmMod(object, ...)
}
\arguments{
\item{object}{A fitted model with class \code{armmMod}.}

\item{...}{Ignored.}
}
\value{
A dataframe of fixed-effects estimates.
}
\description{
Extract fixed-effects estimates from an \code{armmMod} object
}
