% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armmMod_class.R
\name{fixef}
\alias{fixef}
\alias{fixef.armmMod}
\title{Extract fixed-effects estimates.}
\usage{
fixef(object, ...)

\method{fixef}{armmMod}(object, ...)
}
\arguments{
\item{object}{A fitted model with class \code{armmMod}.}

\item{...}{Ignored.}
}
\value{
A dataframe of fixed-effects estimates.
}
\description{
Extract fixed-effects estimates.
}
\section{Methods (by class)}{
\itemize{
\item \code{armmMod}: Extract fixed-effects estimates from an \code{armmMod} object
}}

