% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armmMod_class.R
\name{residuals.communityPGLMM}
\alias{residuals.communityPGLMM}
\title{Residuals of \code{armmMod} objects}
\usage{
\method{residuals}{communityPGLMM}(object, type = "response", ...)
}
\arguments{
\item{object}{A fitted model with class \code{armmMod}.}

\item{type}{Type of residuals, currently only \code{"response"} is programmed.}

\item{\dots}{Additional arguments, ignored for method compatibility.}
}
\value{
A vector of residuals.
}
\description{
Getting different types of residuals for \code{armmMod} objects.
}
